use sqrtf

+++ src/Sphere.cpp
@@ -8,7 +8,7 @@ bool rayIntersectSphere(const Sphere& sphere, const Vec3f& nRay, float* distance
     if (discriminant < 0) { return false; }
 
     if (distance) {
-        *distance = (discriminant > 0) ? (dc - sqrt(discriminant)) : (dc);
+        *distance = (discriminant > 0) ? (dc - sqrtf(discriminant)) : (dc);
     }
     return true;
 }
@@ -22,7 +22,7 @@ bool rayIntersectSphere(const Sphere &sphere, const Vec3f &nRay, const Vec3f &ra
     if (discriminant < 0) { return false; }
 
     if (distance) {
-        *distance = (discriminant > 0) ? (dc - sqrt(discriminant)) : (dc);
+        *distance = (discriminant > 0) ? (dc - sqrtf(discriminant)) : (dc);
     }
     return true;
 }
